Move BaseContext Array methods to NumPy module
