+++ numba/core/base.py
@@ -1017,17 +1017,6 @@ class BaseContext(object):
         """
         return self._make_helper(builder, typ, ref=ref, kind='data')
 
-    def make_array(self, typ):
-        from numba.np import arrayobj
-        return arrayobj.make_array(typ)
-
-    def populate_array(self, arr, **kwargs):
-        """
-        Populate array structure.
-        """
-        from numba.np import arrayobj
-        return arrayobj.populate_array(arr, **kwargs)
-
     def make_complex(self, builder, typ, value=None):
         """
         Get a helper object to access the given complex numbers' members.
@@ -1043,57 +1032,6 @@ class BaseContext(object):
         for i, val in enumerate(values):
             tup = builder.insert_value(tup, val, i)
         return tup
-
-    def make_constant_array(self, builder, typ, ary):
-        """
-        Create an array structure reifying the given constant array.
-        A low-level contiguous array constant is created in the LLVM IR.
-        """
-        datatype = self.get_data_type(typ.dtype)
-        # don't freeze ary of non-contig or bigger than 1MB
-        size_limit = 10**6
-
-        if (self.allow_dynamic_globals and
-                (typ.layout not in 'FC' or ary.nbytes > size_limit)):
-            # get pointer from the ary
-            dataptr = ary.ctypes.data
-            data = self.add_dynamic_addr(builder, dataptr, info=str(type(dataptr)))
-            rt_addr = self.add_dynamic_addr(builder, id(ary), info=str(type(ary)))
-        else:
-            # Handle data: reify the flattened array in "C" or "F" order as a
-            # global array of bytes.
-            flat = ary.flatten(order=typ.layout)
-            # Note: we use `bytearray(flat.data)` instead of `bytearray(flat)` to
-            #       workaround issue #1850 which is due to numpy issue #3147
-            consts = cgutils.create_constant_array(llvmir.IntType(8), bytearray(flat.data))
-            data = cgutils.global_constant(builder, ".const.array.data", consts)
-            # Ensure correct data alignment (issue #1933)
-            data.align = self.get_abi_alignment(datatype)
-            # No reference to parent ndarray
-            rt_addr = None
-
-        # Handle shape
-        llintp = self.get_value_type(types.intp)
-        shapevals = [self.get_constant(types.intp, s) for s in ary.shape]
-        cshape = cgutils.create_constant_array(llintp, shapevals)
-
-        # Handle strides
-        stridevals = [self.get_constant(types.intp, s) for s in ary.strides]
-        cstrides = cgutils.create_constant_array(llintp, stridevals)
-
-        # Create array structure
-        cary = self.make_array(typ)(self, builder)
-
-        intp_itemsize = self.get_constant(types.intp, ary.dtype.itemsize)
-        self.populate_array(cary,
-                            data=builder.bitcast(data, cary.data.type),
-                            shape=cshape,
-                            strides=cstrides,
-                            itemsize=intp_itemsize,
-                            parent=rt_addr,
-                            meminfo=None)
-
-        return cary._getvalue()
 
     def add_dynamic_addr(self, builder, intaddr, info):
         """
